2024.3.9 更新内容： 1. 增加了先前提交未上传的上位机环形缓冲区类 2. 上位机界面控件的命名更新 3. 下位机运行逻辑初版完成（较多工作），待测试 4. 下位机添加一条LIN总线，并添加了相应的驱动，更改了API接口，相应做出了调整进行适应，待测试

+++ Vehicle-thermal-management/app/LIN/LIN_device_control.h
@@ -22,7 +22,7 @@
 /************************************ 压缩机通讯 ************************************/
 
 /* 更改压缩机的转速，变相等于开启压缩机 */
-uint8_t Compressor_Set_Speed(uint16_t speed, uint16_t limit_power);
+uint8_t Compressor_Set_Speed(uint8_t speed, uint8_t limit_power);
 /* 关闭压缩机 */
 uint8_t Compressor_Shutdown(void);
 /* 获取压缩机状态 */
